feat: migrate appView.js to TS
