+++ migration-newip-to-ts/src/components/view/news/news.ts
@@ -2,7 +2,7 @@ import './news.css';
 import { NewsData } from '../../../types';
 
 class News {
-    protected draw(data: readonly NewsData[]): void {
+    public draw(data: readonly NewsData[]): void {
         const news: readonly NewsData[] = data.length >= 10 ? data.filter((_item, idx) => idx < 10) : data;
 
         const fragment: DocumentFragment = document.createDocumentFragment();
